requirements.txt
langgraph
playwright
pyyaml
python-dotenv
neo4j
anthropic
sendgrid
email-validator
bs4
requests
pyhunter
tenacity